Rename arrayAndAverageManyFast to averageManyFast
